# public/openapi.yaml
openapi: 3.0.1
info:
  title: ChatGPT GIS API
  description: API for querying Metro Nashville and TN parcel data
  version: "1.0.0"

servers:
  - url: "https://chatgpt-gis-vercel.vercel.app"
    description: Production API server

tags:
  - name: Health
    description: Endpoints for checking API status
  - name: Metro
    description: Endpoints for querying Metro Nashville parcel data
  - name: Statewide
    description: Endpoints for querying Tennessee-wide parcel data
  - name: Utilities
    description: Utility and placeholder endpoints

paths:
  /api/ping:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: Returns a simple pong response to confirm the API is running.
      operationId: getPing
      responses:
        '200':
          description: Pong response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: pong

  /api/parcel-summary:
    get:
      tags:
        - Statewide
      summary: Get normalized parcel summaries for comma-separated APNs
      operationId: getParcelSummaries
      parameters:
        - name: pins
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated APN(s), e.g. `03200001000,03200001001`
        - name: source
          in: query
          required: false
          schema:
            type: string
            enum: [metro, tn]
          description: Source dataset (metro or tn)
      responses:
        '200':
          description: Normalized parcel summaries
          content:
            application/json:
              schema:
                type: object
                properties:
                  source:
                    type: string
                    example: metro
                  count:
                    type: integer
                    example: 1
                  features:
                    type: array
                    items:
                      type: object
                      properties:
                        apn:
                          type: string
                          example: "03200001000"
                        owner:
                          type: string
                          example: "John Doe"
                        prop_addr:
                          type: string
                          example: "123 Main St"
                        acres:
                          type: number
                          example: 1.25
                        assessed_total:
                          type: number
                          example: 250000
                        land_assessed:
                          type: number
                          example: 100000
                        impr_assessed:
                          type: number
                          example: 150000
                        geometry:
                          type: object
                          description: Esri JSON geometry object
                        raw:
                          type: object
                          description: Original feature from upstream API

  /api/metro:
    get:
      tags:
        - Metro
      summary: Query Metro Nashville parcels by pin or where clause
      operationId: getMetroParcels
      parameters:
        - name: pin
          in: query
          required: false
          schema:
            type: string
        - name: where
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Metro parcel result
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        apn:
                          type: string
                          example: "03200001000"
                        zoning:
                          type: string
                          example: "R10"
                        owner:
                          type: string
                          example: "Jane Smith"
                        where:
                          type: string
                          example: "zoning='R10'"

  /api/tn:
    get:
      tags:
        - Statewide
      summary: Query TN parcels by pin or where clause
      operationId: getTnParcels
      parameters:
        - name: pin
          in: query
          required: false
          schema:
            type: string
        - name: where
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: TN parcel result
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        apn:
                          type: string
                          example: "09400002000"
                        county:
                          type: string
                          example: "Davidson"
                        owner:
                          type: string
                          example: "Acme Corp"
                        where:
                          type: string
                          example: "county='Davidson'"

  /api/parcel:
    get:
      tags:
        - Utilities
      summary: Placeholder parcel endpoint
      description: Returns a placeholder message for the parcel route.
      operationId: getParcelPlaceholder
      responses:
        '200':
          description: Placeholder response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Parcel endpoint placeholder"